refactor: changed poi_id
Changed poi_id to fit with flake8 linting. Changed the way sklearn modules were imported. Placed at the top of the file and explicitly stated which module a method is called from. e.g. naive_bayes.GaussianNB(). This is to make it easier to read where each of these have come from.