fix: now running on python 3
Had to modify and fix some errors. The file final_project_dataset.pkl was not being recognised. Added an os.path.abspath solution for this.

Using python 3 pickle was throwing up a TypeError about expecting bytes instead of strings. Changed to "rb" and "wb" instead of "r" and "w" respectively.